- Fix character name - Removed ConanArray - Pulled in the Vector3 from monogame.net - Cleanup database - Cleanup ConanStream

+++ src/LibFaolan/Data/Realm.cs
@@ -4,13 +4,13 @@ namespace LibFaolan.Data
 {
     public sealed class Realm
     {
-        public UInt32 Country; //Country: 0 England, 1 France, 2 Spain, 3 German
-        public UInt32 FullStatus; // server full status 1 = full
+        public byte Country; //Country: 0 England, 1 France, 2 Spain, 3 German
+        public byte FullStatus; // server full status 1 = full
         public string IpAddress;
-        public UInt32 LoadStatus; //load status -- 0 = medium load 1 = medium load 2 = heavy load 3 = full
-        public UInt32 OnlineStatus; // onlineStatus - 2 = online, other num = offline
-        public UInt32 Port;
-        public UInt32 RealmId;
+        public byte LoadStatus; //load status -- 0 = medium load 1 = medium load 2 = heavy load 3 = full
+        public byte OnlineStatus; // onlineStatus - 2 = online, other num = offline
+        public UInt16 Port;
+        public byte RealmId;
         public string RealmName;
         public UInt16 RealmType; // realmtype - 0,1 = PvE | 256,257 = PvP | 1081 = PvP by faction, probably a bitmask
     }
